requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
contextlib2==0.5.4
defusedxml==0.4.1
Django==1.10.3
django-bootstrap3==7.1.0
djangorestframework==3.5.3
kombu==3.0.37
oauthlib==2.0.1
Pillow==3.4.2
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.7
raven==5.32.0
redis==2.10.5
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
